Add rudimentary deserialization of extension lists.
Add ability to deserialize an extension list.  Note that we really only deserialize the length of the extension array, because schema of following extensions is unknown. But these schemas may be known at the calling level, or else a non-zero list size can be used to flag the existence of unparsable op data.

Use of the extension list is exemplified in the deserialization of the Transfer op. Future commits will add handling to other ops.