Change semantics of _prepro callback return types to resolve compiler warnings.
Callback typedef bagl_element_callback_t specifies a const bagl_element_t * as the return val, but these functions were defined to return the (binary equivalent) unsigned int, and return statements were using that as a signaling mechanism of some sort.  So, I fixed the return type, but now have to cast the return values to avoid warnings. Not certain this is best way to clean this up, but it DOES resolve the compiler warnings.